Create MockOrderManager class
